requirements.txt
tweepy==4.14.0
python-dotenv==0.19.2
APScheduler==3.9.1
pandas==1.4.0
tensorflow-cpu==2.8.0
python-magic==0.4.15
numpy==1.22.1
tqdm==4.63.0
boto3==1.21.22
atproto==0.0.61